% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlunitjars-package.R
\docType{package}
\name{htmlunitjars}
\alias{htmlunitjars}
\alias{htmlunitjars-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/htmlunitjars}
\item BugReports: \url{https://gitlab.com/hrbrmstr/htmlunitjars/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
