% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlunitjars-package.R
\docType{package}
\name{htmlunitjars}
\alias{htmlunitjars}
\alias{htmlunitjars-package}
\title{Contents of the 'HtmlUnit' & supporting Java archives \url{http://htmlunit.sourceforge.net/}}
\description{
Library JavaDoc: \url{http://htmlunit.sourceforge.net/apidocs/index.html}
}
\details{
HtmlUnit is a "GUI-Less browser for Java programs". It models HTML documents and provides an API that allows you to invoke pages, fill out forms, click links, etc... just like you do in your "normal" browser.

It has fairly good JavaScript support (which is constantly improving) and is able to work even with quite complex AJAX libraries, simulating Chrome, Firefox or Internet Explorer depending on the configuration used.

It is typically used for testing purposes or to retrieve information from web sites.

HtmlUnit is not a generic unit testing framework. It is specifically a way to simulate a browser.
}
\author{
Bob Rudis (bob@rud.is)
}
